+++ src/server/services/allRepos/index.ts
@@ -1,10 +1,16 @@
+import { SERVICE_Response__BASE } from '..'
+
 import { PAGE_SIZE } from '@/server/lib/gh-api'
 import { GH_API_Call__countRepos, GH_API_Call__listRepos } from '@/server/lib/gh-api/repos'
 import { chunkArr } from '@/server/utils/chunkArr'
-import { SHARED_APIFields__CountRepos, SHARED_APIFields__ListRepos, SHARED_Data__Repo } from '@/shared/models'
+import { SHARED_Data__Repo } from '@/shared/models/models/Repos'
+
+interface SERVICE_Response__listRepos extends SERVICE_Response__BASE {
+    repos?: SHARED_Data__Repo[]
+}
 
 // https://docs.github.com/en/rest/repos/repos?apiVersion=2022-11-28#list-repositories-for-the-authenticated-user
-export const listRepos = async (accessToken: string): Promise<SHARED_APIFields__ListRepos> => {
+export const SERVICE_Call__listRepos = async (accessToken: string): Promise<SERVICE_Response__listRepos> => {
     const { success, error, repos } = await GH_API_Call__listRepos(accessToken)
 
     if (!success || repos === undefined) {
@@ -16,7 +22,11 @@ export const listRepos = async (accessToken: string): Promise<SHARED_APIFields__
 
 /* ======== count/fetch all repos ======== */
 
-export const countAllRepos = async (accessToken: string): Promise<SHARED_APIFields__CountRepos> => {
+interface SERVICE_Response__countAllRepos extends SERVICE_Response__BASE {
+    numRepos?: number
+}
+
+export const countAllRepos = async (accessToken: string): Promise<SERVICE_Response__countAllRepos> => {
     const { success, error, numRepos } = await GH_API_Call__countRepos(accessToken)
 
     if (!success || numRepos === undefined) {
@@ -26,7 +36,9 @@ export const countAllRepos = async (accessToken: string): Promise<SHARED_APIFiel
     return { numRepos: numRepos, success: true }
 }
 
-export const listAllRepos = async (accessToken: string): Promise<SHARED_APIFields__ListRepos> => {
+type SERVICE_Response__listAllRepos = SERVICE_Response__listRepos
+
+export const listAllRepos = async (accessToken: string): Promise<SERVICE_Response__listAllRepos> => {
     const { success: countSuccess, error: countError, numRepos } = await countAllRepos(accessToken)
     // console.log(numRepos)
 
@@ -34,7 +46,7 @@ export const listAllRepos = async (accessToken: string): Promise<SHARED_APIField
         return { repos: undefined, success: false, error: countError }
     }
 
-    const repoResPromises: Promise<SHARED_APIFields__ListRepos>[] = []
+    const repoResPromises: Promise<SERVICE_Response__listRepos>[] = []
     for (let page = 1; (page - 1) * PAGE_SIZE < numRepos; page++) {
         repoResPromises.push(
             GH_API_Call__listRepos(accessToken, {
@@ -47,7 +59,7 @@ export const listAllRepos = async (accessToken: string): Promise<SHARED_APIField
     const chunkedRepoResPromises = chunkArr(repoResPromises)
 
     try {
-        const repoResArr: SHARED_APIFields__ListRepos[] = []
+        const repoResArr: SERVICE_Response__listRepos[] = []
         for (let i = 0; i < chunkedRepoResPromises.length; i++) {
             repoResArr.push(...(await Promise.all(chunkedRepoResPromises[i])))
         }
